% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec-package.R
\docType{package}
\name{sec-package}
\alias{sec}
\alias{sec-package}
\title{sec: Securities and Exchange Commission APIs}
\description{
Search for companies, and download and parse their Securities and Exchange Commission (SEC) filings.
}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Authors:
\itemize{
  \item Biocomplexity Institute [copyright holder]
}

}
\keyword{internal}
