% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_company.R
\name{sec_companies}
\alias{sec_companies}
\title{Download Company Information}
\usage{
sec_companies(ciks, type = "submissions", out = tempdir(), load = TRUE,
  overwrite = FALSE, user = Sys.getenv("SEC_USER"), verbose = TRUE)
}
\arguments{
\item{ciks}{A vector of Central Index Keys, or a \code{data.frame} with a \code{ciks}
column.}

\item{type}{Type of information to collect, between \code{submissions}, which includes things
like Standard Industrial Classification, ticker, and previous names, as well as information
about recent filings, and \code{facts}, including financial}

\item{out}{Path to a directory in which to save filings. Each filing will be saved
within a subdirectory: \code{{out}/{cik}/{SEC Accession Number}}.}

\item{load}{Logical; if \code{TRUE}, will read in the downloaded info file.}

\item{overwrite}{Logical; if \code{TRUE}, download the info file even if it already exists.}

\item{user}{User agent string (e.g., \code{"Company Name AdminContact@example.com"});
defaults to the \code{SEC_USER} environment variable, which can be set with
\code{\link{Sys.setenv}} (e.g., \code{Sys.setenv(SEC_USER = "...")}).}

\item{verbose}{Logical; if \code{FALSE}, will not show status updates.}
}
\value{
An invisible list with named entries for each successful CIK if \code{load} is \code{TRUE},
or a vector of HTTP status codes otherwise.
}
\description{
Retrieve information about companies.
}
\examples{
\dontrun{

# retrieve information about Pepsico
pepsi <- sec_companies("0000077476")
pepsi$`0000077476`[1:10]

# retrieve financial information about Pepsico
pepsi_facts <- sec_companies("0000077476", "facts")
pepsi_facts$`0000077476`$facts$`us-gaap`$Revenues
}
}
\seealso{
Search for filings with \code{\link{sec_search}}.
}
