% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_download.R
\name{sec_download}
\alias{sec_download}
\title{Download Filings}
\usage{
sec_download(ciks, out, ids = NULL, complete = FALSE,
  user = Sys.getenv("SEC_USER"), verbose = TRUE)
}
\arguments{
\item{ciks}{A vector of Central Index Keys, or a \code{data.frame} with a \code{ciks}
column.}

\item{out}{Path to a directory in which to save filings. Each filing will be saved
within a subdirectory: \code{{out}/{cik}/{SEC Accession Number}}.}

\item{ids}{A vector of document IDs. These are combinations of SEC Accession Numbers
and file names, separated by a colon (e.g., \code{"0001564590-22-032043:gs-424b2.htm"}).
If the file name is not included, this will default to the complete submission text file
(e.g., \code{"0001564590-22-032043.txt"}).}

\item{complete}{Logical; if \code{TRUE}, will download the complete submission text file
rather than another files specified after a colon in \code{ids}.}

\item{user}{User agent string (e.g., \code{"Company Name AdminContact@example.com"});
defaults to the \code{SEC_USER} environment variable, which can be set with
\code{\link{Sys.setenv}} (e.g., \code{Sys.setenv(SEC_USER = "...")}).}

\item{verbose}{Logical; if \code{FALSE}, will not show status updates.}
}
\value{
An invisible vector of HTTP status codes.
}
\description{
Download SEC filings based on Central Index Key (CIK) and document ID (as returned
from \code{\link{sec_search}}.
}
\examples{
\dontrun{

# search for filings based on SIC
forms <- sec_search(SICs = 2080, forms = "10-K", limit = 100)

# download some of those filings
sec_download(forms[1:3, ], out = tempdir())
}
}
\seealso{
Search for filings with \code{\link{sec_search}}.
}
