% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_search.R
\name{sec_search}
\alias{sec_search}
\title{Search for Filings}
\usage{
sec_search(query = NULL, entity = NULL, CIKs = NULL, SICs = NULL,
  forms = NULL, date_to = NULL, date_from = "2001-01-01",
  outFile = NULL, overwrite = TRUE, start = 0, limit = Inf,
  verbose = TRUE)
}
\arguments{
\item{query}{A character with words or phrases to search for within the form
(e.g., \code{grape "fruit juice"}).}

\item{entity}{A character with a name, ticker, or Central Index Key.}

\item{CIKs}{A vector of Central Index Keys (e.g., \code{c(0000317540, 0000077476)}).}

\item{SICs}{A vector of Standard Industrial Classification codes (e.g., \code{2080}).}

\item{forms}{A vector of forms to include (e.g., \code{"10-K"}).}

\item{date_to}{Latest date to include results from (e.g., \code{"2019-01-01"}); defaults to current.}

\item{date_from}{Earliest date to include results from; defaults to the earliest available:
\code{"2001-01-01"}.}

\item{outFile}{Path of a file to write results to.}

\item{overwrite}{Logical; if \code{FALSE}, will read in an existing results file rather than
executing the search.}

\item{start}{Number of results to skip.}

\item{limit}{Maximum number of results to return, in steps of 100.}

\item{verbose}{Logical; if \code{FALSE}, will no display status messages.}
}
\value{
A \code{data.frame} of results.
}
\description{
Search for forms by keywords, Central Index Keys, and/or Standard Industrial Classification
codes, filed after 2001.
}
\examples{
\dontrun{

# search for 10-K filings by beverage companies
results <- sec_search(SICs = 2080, forms = "10-K", limit = 100)
}
}
\seealso{
To retrieve a simpler list of companies, see \code{\link{sec_search_companies}}.
}
