% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_search_companies.R
\name{sec_search_companies}
\alias{sec_search_companies}
\title{Search for Companies (Legacy)}
\usage{
sec_search_companies(SIC, start = 0, limit = Inf,
  user = Sys.getenv("SEC_USER"), verbose = TRUE)
}
\arguments{
\item{SIC}{Standard Industrial Classification code (e.g., \code{2080}).}

\item{start}{Number of results to skip.}

\item{limit}{Maximum number of results to return, in steps of 100.}

\item{user}{User agent string (e.g., \code{"Company Name AdminContact@example.com"});
defaults to the \code{SEC_USER} environment variable, which can be set with
\code{\link{Sys.setenv}} (e.g., \code{Sys.setenv(SEC_USER = "...")}).}

\item{verbose}{Logical; if \code{FALSE}, will not show status updates.}
}
\value{
A \code{data.frame} of results, with columns for \code{CIK} (Central Index Key),
\code{Company}, and \code{Location}.
}
\description{
Search for companies based on Standard Industrial Classification.
}
\examples{
\dontrun{

# get a list of beverage companies
sec_search_companies(2080)
}
}
\seealso{
For more complete results, see \code{\link{sec_search}}.
}
